Change global forms variable to a function - so default form values can't be modified. Buildbot now automatically generates getForms function change isvalid functions to use string slice so function signatures are more consistent Store form expiry date directly on form so sessionInfo struct is no longer needed change form actions integer to uint8 comment out All page

+++ !/pClub.go
@@ -29,7 +29,7 @@ func clubs(w http.ResponseWriter, r *http.Request) {
 	templater(w, page{
 		Title: "Clubs",
 		Data: M{
-			"NewClub":   getSession(w, r, []int{clubNew}),
+			"NewClub":   getSession(w, r, []uint8{clubNew}),
 			"ListClubs": listClubs,
 		},
 	})
